+++ lib/app.js
@@ -4,20 +4,18 @@ const express = require('express');
 const favicon = require('serve-favicon');
 const bodyParser = require('body-parser');
 const cookieParser = require('cookie-parser');
-const validator = require('express-validator');
-const responseTime = require('response-time');
-const logger = require('./logger')('app.js');
 const routes = require('./routes');
 const path = require('path');
 const errorHandler = require('errorhandler');
-
-const { REQUEST_TIME } = require('./events');
+const unleashSession = require('./middleware/session');
+const responseTime = require('./middleware/response-time');
+const requestLogger = require('./middleware/request-logger');
+const validator = require('./middleware/validator');
 
 module.exports = function(config) {
     const app = express();
 
     const baseUriPath = config.baseUriPath || '';
-    const publicFolder = config.publicFolder;
 
     app.set('trust proxy');
     app.disable('x-powered-by');
@@ -29,42 +27,15 @@ module.exports = function(config) {
     }
 
     app.use(cookieParser());
-
-    if (publicFolder) {
-        app.use(favicon(path.join(publicFolder, 'favicon.ico')));
-    }
-
-    app.use(
-        responseTime((req, res, time) => {
-            const timingInfo = {
-                path: req.baseUrl,
-                method: req.method,
-                statusCode: res.statusCode,
-                time,
-            };
-            config.eventBus.emit(REQUEST_TIME, timingInfo);
-        })
-    );
-
-    app.use(
-        validator({
-            customValidators: {
-                isUrlFirendlyName: input => encodeURIComponent(input) === input,
-            },
-        })
-    );
-
-    if (publicFolder) {
-        app.use(baseUriPath, express.static(publicFolder));
-    }
-
     app.use(bodyParser.json({ strict: false }));
+    app.use(unleashSession(config));
+    app.use(responseTime(config));
+    app.use(requestLogger(config));
+    app.use(validator(config));
 
-    if (config.enableRequestLogger) {
-        app.use((req, res, next) => {
-            next();
-            logger.info(`${res.statusCode} ${req.method} ${req.baseUrl}`);
-        });
+    if (config.publicFolder) {
+        app.use(favicon(path.join(config.publicFolder, 'favicon.ico')));
+        app.use(baseUriPath, express.static(config.publicFolder));
     }
 
     if (typeof config.preRouterHook === 'function') {
