--- lib/middleware/session-middleware.js
@@ -1,8 +1,0 @@
-'use strict';
-
-const cookieSession = require('cookie-session');
-
-module.exports = config => {
-    config.a = 1;
-    return cookieSession(config.field);
-};
