+++ lib/middleware/session.js
@@ -1,0 +1,11 @@
+'use strict';
+
+const cookieSession = require('cookie-session');
+
+module.exports = function(config) {
+    return cookieSession({
+        name: 'unleash-session',
+        keys: [config.secret],
+        maxAge: config.sessionAge,
+    });
+};
