+++ lib/options.js
@@ -3,6 +3,7 @@
 const { publicFolder } = require('unleash-frontend');
 
 const isDev = () => process.env.NODE_ENV === 'development';
+const THIRTY_DAYS = 30 * 24 * 60 * 60 * 1000;
 
 const DEFAULT_OPTIONS = {
     databaseUrl: process.env.DATABASE_URL,
@@ -12,6 +13,8 @@ const DEFAULT_OPTIONS = {
     enableLegacyRoutes: true,
     publicFolder,
     enableRequestLogger: isDev(),
+    secret: 'UNLEASH-SECRET',
+    sessionAge: THIRTY_DAYS,
 };
 
 module.exports = {
