+++ lib/routes/admin-api/index.js
@@ -24,9 +24,7 @@ exports.apiDef = apiDef;
 exports.router = config => {
     const router = Router();
 
-    router.get('/', (req, res) => {
-        res.json(apiDef);
-    });
+    router.get('/', (req, res) => res.json(apiDef));
 
     router.use('/features', features.router(config));
     router.use('/archive', featureArchive.router(config));
